requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
curl_cffi==0.13.0
Deprecated==1.3.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.135.1
feedparser==6.0.12
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==4.1.1
fredapi==0.5.2
frozendict==2.4.7
gunicorn==25.1.0
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
multitasking==0.0.12
numpy==2.2.6
openpyxl==3.1.5
ordered-set==4.1.0
packaging==26.0
pandas==2.3.3
peewee==4.0.0
platformdirs==4.9.2
protobuf==7.34.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sgmllib3k==1.0.0
xlrd==2.0.1
six==1.17.0
slowapi==0.1.9
soupsieve==2.8.3
starlette==0.52.1
stripe==14.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.6
wrapt==2.1.1
yfinance==1.2.0
